edge-server/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
llama-cpp-python
faster-whisper
huggingface_hub
Pillow
pytesseract
requests
